Split ONNX framework into multiple files
